requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.3
autopep8==1.4.4
awsebcli==3.17.0
billiard==3.6.1.0
blessed==1.16.1
boto==2.49.0
boto3==1.10.50
botocore==1.13.50
cached-property==1.5.1
celery==4.4.0
cement==2.8.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.3.9
Django==3.0.2
django-allow-cidr==0.3.1
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-fontawesome==1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-static-fontawesome==5.10.2
django-storages==1.8
django-timezone-field==4.0
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
filelock==3.0.12
future==0.16.0
gevent==1.4.0
greenlet==0.4.15
idna==2.7
importlib-metadata==1.3.0
jmespath==0.9.4
jsonschema==2.6.0
kombu==4.6.7
more-itertools==8.0.2
netaddr==0.7.19
packaging==20.0
pathspec==0.5.9
Pillow==6.2.1
pluggy==0.13.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pycurl==7.43.0.3
pyparsing==2.4.6
python-crontab==2.4.0
python-dateutil==2.8.0
pytz==2019.3
PyYAML==3.13
redis==3.3.11
requests==2.20.1
s3transfer==0.2.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
toml==0.10.0
urllib3==1.24.3
vine==1.3.0
virtualenv==16.7.9
wcwidth==0.1.8
websocket-client==0.57.0
zipp==0.6.0
